급여 이모저모 수정

+++ src/main/java/kr/co/groovy/vo/PaystubVO.java
@@ -1,6 +1,5 @@
 package kr.co.groovy.vo;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -16,7 +15,7 @@ public class PaystubVO {
     private String salaryEmplId; // salary, salary_dtsmt
     private String salaryEmplNm; // salary, salary_dtsmt
     private int salaryBslry; // 통상임금 | salary
-    private double salaryOvtimeAllwnc; // 시간외수당 | salary
+    private int salaryOvtimeAllwnc; // 시간외수당 | salary
     private int salaryDtsmtPymntTotamt; // 지급액계 | salary_dtsmt
     private int salaryDtsmtSisNp; // 국민연금 | salary_dtsmt
     private int salaryDtsmtSisHi; // 건강보험 | salary_dtsmt
